backups_repair_20251005_230417/src_before_repair/jeffrey/core/neuralbus/ffi_cdata.py
"""
Pont FFI Zéro-Copie via Arrow C Data Interface
VERSION CORRIGÉE : Mode stub sans charger de lib inexistante
"""

import ctypes
import logging
import os
from typing import Any

import pyarrow as pa

logger = logging.getLogger(__name__)


class ZeroCopyFFIBridge:
    """
    Bridge haute performance Python-Rust via Arrow C Data Interface
    Mode STUB en DEV tant que la lib Rust n'est pas compilée
    """

    def __init__(self):
        self.rust_lib: ctypes.CDLL | None = None
        self.arrow_schemas = {}
        self.mode = os.getenv("SECURITY_MODE", "dev")

    def initialize(self):
        """Initialise le pont FFI (ou reste en stub)"""
        # Définir les schémas
        self._define_schemas()

        # Charger la lib Rust SEULEMENT si elle existe
        if self.mode == "prod":
            try:
                # TODO: Compiler avec maturin d'abord
                lib_path = "./target/release/jeffrey_core.so"
                if os.path.exists(lib_path):
                    self.rust_lib = ctypes.CDLL(lib_path)
                    logger.info("✅ Rust library loaded")
                else:
                    logger.warning("⚠️ Rust library not found, using stub mode")
            except Exception as e:
                logger.warning(f"⚠️ Could not load Rust lib: {e}, using stub mode")
        else:
            logger.info("📝 FFI Bridge in STUB mode (DEV)")

        logger.info("✅ Zero-Copy FFI Bridge initialized")

    def _define_schemas(self):
        """Définit les schémas Arrow pour l'échange"""
        # Schéma pour les événements
        self.arrow_schemas["event"] = pa.schema(
            [
                ("id", pa.string()),
                ("source", pa.string()),
                ("event_type", pa.string()),
                ("severity", pa.float32()),
                ("timestamp", pa.timestamp("us", tz="UTC")),
                ("data", pa.binary()),
            ]
        )

        # Schéma pour les métriques
        self.arrow_schemas["metrics"] = pa.schema(
            [
                ("cpu_usage", pa.float32()),
                ("memory_mb", pa.int32()),
                ("events_per_sec", pa.int32()),
                ("latency_ms", pa.float32()),
            ]
        )

        # Schéma pour threat analysis
        self.arrow_schemas["threat"] = pa.schema(
            [
                ("threat_level", pa.float32()),
                ("anomaly_score", pa.float32()),
                ("confidence", pa.float32()),
                ("action", pa.string()),
            ]
        )

    def export_to_c(self, data: Any, schema_name: str) -> tuple[Any, Any]:
        """
        Exporte des données Python vers le format C (pour Rust)
        En mode stub, retourne des mocks
        """
        if not self.rust_lib:
            # Mode stub - retourne des placeholders
            return (None, None)

        schema = self.arrow_schemas.get(schema_name)
        if not schema:
            raise ValueError(f"Unknown schema: {schema_name}")

        # Convertir en RecordBatch Arrow
        if isinstance(data, dict):
            arrays = []
            for field in schema:
                value = data.get(field.name)
                if value is not None:
                    arrays.append(pa.array([value]))
                else:
                    arrays.append(pa.array([None]))

            batch = pa.RecordBatch.from_arrays(arrays, schema=schema)
        elif isinstance(data, list):
            batch = pa.RecordBatch.from_pylist(data, schema=schema)
        else:
            batch = data

        # Export vers C Data Interface
        # TODO: Implémenter quand Rust est prêt
        # c_array = batch._export_to_c()
        # c_schema = schema._export_to_c()

        return (None, None)  # Stub pour l'instant

    def import_from_c(self, c_array: Any, c_schema: Any) -> pa.RecordBatch | None:
        """
        Importe des données depuis Rust (format C)
        En mode stub, retourne None
        """
        if not self.rust_lib:
            return None

        # TODO: Implémenter quand Rust est prêt
        # batch = pa.RecordBatch._import_from_c(c_array, c_schema)
        # return batch

        return None

    def call_rust_function(self, function_name: str, data: dict, schema_name: str) -> dict:
        """
        Appelle une fonction Rust avec zéro-copie
        En mode stub, retourne un mock
        """
        if not self.rust_lib:
            # Mode stub - retourne un mock
            logger.debug(f"STUB: Would call Rust function {function_name}")
            return {
                "status": "mock",
                "function": function_name,
                "mode": "stub",
                "data_received": bool(data),
            }

        # TODO: Implémenter avec PyO3 quand Rust est prêt
        # c_array, c_schema = self.export_to_c(data, schema_name)
        # result = self.rust_lib.process(c_array, c_schema)
        # result_batch = self.import_from_c(result_array, result_schema)

        return {"status": "processed"}

    def get_status(self) -> dict[str, Any]:
        """Retourne le statut du bridge"""
        return {
            "mode": self.mode,
            "rust_loaded": self.rust_lib is not None,
            "schemas_defined": list(self.arrow_schemas.keys()),
            "stub_mode": self.rust_lib is None,
        }
