requirements-optional.txt
# Dashboard & Visualization
streamlit>=1.25.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Graph & Network Analysis
networkx>=3.0

# Machine Learning
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
numpy>=1.24
scipy>=1.10

# Compression
python-snappy>=0.6.1
lz4>=4.3.0

# Monitoring
prometheus-client>=0.17.0

# Advanced ML (future) - Commented out for now (heavy on Mac)
# torch>=2.0.0  # For GNN later
# torch-geometric>=2.3.0  # For predictive graphs

# Testing
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
pytest-cov>=4.1.0
